fix entrypoint issue
